completed refactoring into blueprints, fixed a bug with the login where duplicate usernames could be made
