+++ blueprints/utils.py
@@ -1,3 +1,4 @@
+from bson import ObjectId
 from pymongo import MongoClient
 import pymongo
 
@@ -101,3 +102,34 @@ def get_id(collection, name):
     if item is not None:
         return item["_id"]
     return None
+
+
+def get_ingredient_and_flavor_list(dataDict):
+    flavor_ids = []
+    for i in dataDict["flavors"]:
+        if any(j["name"] == i for j in flavors):
+            for j in flavors:
+                if j["name"] == i:
+                    flavor_ids.append(ObjectId(j["_id"]["$oid"]))
+        else:
+            flavor_ids.append(
+                ObjectId(add_flavor_return_id(i))
+            )
+
+    ingredient_ids = []
+    for i in dataDict["ingredients"]:
+        if any(j["name"] == i["name"] for j in ingredients):
+            for j in ingredients:
+                if j["name"] == i["name"]:
+                    ingredient_ids = j["_id"]["$oid"]
+        else:
+                ingredient_ids = add_ingredient_return_id(i["name"], i["type"])
+
+        ingredient_ids.append(
+            {
+                "ingredient": ObjectId(ingredients_id),
+                "quantity": i['quantity'],
+                "units": i['units'],
+                "type": i['type']
+            })
+    return [flavor_ids, ingredient_ids]
