+++ templates/addcocktail.html
@@ -6,7 +6,7 @@
             <div>
                 <h4>New Cocktail Details:</h4>
             </div>
-            <div><a class="btn btn-outline-dark" href="{{ url_for('index') }}"><i class="fas fa-times"></i></a></div>
+            <div><a class="btn btn-outline-dark" href="{{ url_for('home.index') }}"><i class="fas fa-times"></i></a></div>
         </div>
         <form >
             <div class="form-group px-4">
