+++ templates/filtered.html
@@ -24,7 +24,7 @@
                         </div>
 
                         <div class="d-flex like-view">
-                            <h3><a href="{{ url_for('view_cocktail', cocktail_id = cocktail._id)}}"><i
+                            <h3><a href="{{ url_for('profile.view_cocktail', cocktail_id = cocktail._id)}}"><i
                                         class="fas fa-eye"></i></a></h3>
                             {% if session['username'] %}
                             {%  if cocktail._id in user.starred_cocktails %}
@@ -118,7 +118,7 @@
                         </div>
                         <div class="mr-2 text-right foot foot-right">
                             <p class="blue"><small> a drink by <a
-                                        href="{{ url_for('view_user_profile', user_id=cocktail.creator._id) }}">{{ cocktail.creator.username }}</a></small>
+                                        href="{{ url_for('profile.view_user_profile', user_id=cocktail.creator._id) }}">{{ cocktail.creator.username }}</a></small>
                             </p>
                             <p class="blue"><small class="timestamp">{{ cocktail.created_at }}</small></p>
                         </div>
