+++ templates/index.html
@@ -11,7 +11,7 @@
                 </div>
                 <div class="card-body">
                     {% if session['username'] %}
-                        <a class="waves-effect waves-light btn modal-trigger" href="{{ url_for('new_drink') }}"><i
+                        <a class="waves-effect waves-light btn modal-trigger" href="{{ url_for('add_drink.new_drink') }}"><i
                             class="fas fa-cocktail"></i>New Drink</a>
                     {% else %}
                         <a class="waves-effect waves-light btn modal-trigger" data-toggle="modal" data-target="#sign-up"><i class="fas fa-cocktail"></i>Join
@@ -34,7 +34,7 @@
                             </div>
                            
                             <div class="d-flex like-view">
-                                <h3><a href="{{ url_for('view_cocktail', cocktail_id = cocktail._id)}}"><i class="fas fa-eye"></i></a></h3>
+                                <h3><a href="{{ url_for('profiles.view_cocktail', cocktail_id = cocktail._id)}}"><i class="fas fa-eye"></i></a></h3>
                                 {% if session['username'] %}
                                 {%  if cocktail._id in user.starred_cocktails %}
                                     <h3 class="cocktails"><i class="fas fa-star" data-id="{{cocktail._id}}" data-user="{{session['_id']}}" 
@@ -119,7 +119,7 @@
                                 </h5>
                             </div>
                             <div class="mr-2 text-right foot foot-right">
-                                <p class="blue"><small> a drink by <a href="{{ url_for('view_user_profile', user_id=cocktail.creator._id) }}">{{ cocktail.creator.username }}</a></small></p>
+                                <p class="blue"><small> a drink by <a href="{{ url_for('profiles.view_user_profile', user_id=cocktail.creator._id) }}">{{ cocktail.creator.username }}</a></small></p>
                                 <p class="blue"><small class="timestamp">{{ cocktail.created_at }}</small></p>
                             </div>
                         </div>
@@ -136,7 +136,7 @@
                     </div>
                     <div class="card-content center">
                          {% if session['username'] %}
-                            <a class="waves-effect waves-light btn modal-trigger" href="{{ url_for('new_drink') }}"><i class="fas fa-cocktail"></i>New Drink</a>
+                            <a class="waves-effect waves-light btn modal-trigger" href="{{ url_for('add_drink.new_drink') }}"><i class="fas fa-cocktail"></i>New Drink</a>
                          {% else %}
                             <a class="waves-effect waves-light btn modal-trigger" data-toggle="modal" data-target="#sign-up"><i class="fas fa-cocktail"></i>Join Today</a>
                          {% endif %}
