+++ templates/viewcocktail.html
@@ -6,7 +6,7 @@
             <div>
                 <h3> {{ cocktail.name }}</h3>
             </div>
-            <div><a class="btn btn-outline-dark" href="{{ url_for('index') }}"><i class="fas fa-times"></i></a></div>
+            <div><a class="btn btn-outline-dark" href="{{ url_for('home.index') }}"><i class="fas fa-times"></i></a></div>
         </div>
         <div class="row d-flex px-4 justify-content-between">
             <div class="col-12 col-lg-6">
